+++ internal/adapter/Executer.go
@@ -1,8 +1,11 @@
 package adapter
 
+import "github.com/fvsystem/gomark/internal/application/shared"
+
 type Executer interface {
 	Execute(
 		testCreator TestCreator,
 		requester Requester,
+		eventEmitter shared.EventEmitter,
 	)
 }