+++ internal/infrastructure/cli/cli.go
@@ -2,11 +2,16 @@ package cli
 
 import (
 	"github.com/fvsystem/gomark/internal/adapter"
+	"github.com/fvsystem/gomark/internal/application/shared"
 )
 
 type CliExecuter struct{}
 
-func (c *CliExecuter) Execute(createTest adapter.TestCreator, requester adapter.Requester) {
+func (c *CliExecuter) Execute(
+	createTest adapter.TestCreator,
+	requester adapter.Requester,
+	eventEmitter shared.EventEmitter,
+) {
 
-	createTest.CreateTest(requester)
+	createTest.CreateTest(requester, 2, eventEmitter)
 }