+++ internal/infrastructure/cli/cli_test.go
@@ -2,6 +2,7 @@ package cli
 
 import (
 	"github.com/fvsystem/gomark/internal/adapter"
+	"github.com/fvsystem/gomark/internal/application/shared"
 
 	"testing"
 )
@@ -10,9 +11,11 @@ type spyCreateTest struct {
 	called bool
 }
 
-func (n *spyCreateTest) CreateTest(requester adapter.Requester) adapter.TestResult {
+func (n *spyCreateTest) CreateTest(requester adapter.Requester,
+	numberOfConnections int,
+	eventEmitter shared.EventEmitter,
+) {
 	n.called = true
-	return adapter.TestResult{}
 }
 
 type fakeRequester struct{}
@@ -21,11 +24,26 @@ func (n *fakeRequester) Get(url string) (adapter.Response, error) {
 	return adapter.Response{}, nil
 }
 
+type FakeEventEmitter struct{}
+
+func (e *FakeEventEmitter) EmitEvent(event shared.Event) {
+
+}
+
+func (e *FakeEventEmitter) AddListener(event string, listener func(event shared.Event)) {
+
+}
+
+func (e *FakeEventEmitter) RemoveAllListeners(event string) {
+
+}
+
 func TestExecute(t *testing.T) {
 	var createTest = &spyCreateTest{called: false}
 	var cliExecuter adapter.Executer = &CliExecuter{}
+	var eventEmitter = &FakeEventEmitter{}
 
-	cliExecuter.Execute(createTest, &fakeRequester{})
+	cliExecuter.Execute(createTest, &fakeRequester{}, eventEmitter)
 
 	if !createTest.called {
 		t.Errorf("Expected %t, got %t", true, createTest.called)